requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
django-redis
hiredis
msgpack-python
